[refactor] ProjectCompany 패키지 리팩토링 변경 (대문자 -> 소문자 시작)
