+++ module-infra/src/main/java/com/welcommu/moduleinfra/project/ProjectCompanyRepository.java
@@ -1,6 +1,6 @@
 package com.welcommu.moduleinfra.project;
 
-import com.welcommu.moduledomain.ProjectCompany.ProjectCompany;
+import com.welcommu.moduledomain.projectCompany.ProjectCompany;
 import com.welcommu.moduledomain.project.Project;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
